+++ R/gg_theme.R
@@ -1,23 +1,23 @@
 #' @title Quick theme for a ggplot.
 #'
 #' @description Quick theme for a ggplot visualisation.
-#' @param font The font for all text to use. Defaults to "".
-#' @param title_font The font for the title. If NULL, inherits from font argument.
-#' @param title_pal The colour palette for the title font. Defaults to "#000000".
-#' @param title_size The size of the title font. Defaults to 11.
-#' @param title_style The style of the title font. Defaults to "bold".
-#' @param subtitle_font The font for the subtitle. If NULL, inherits from font argument.
-#' @param subtitle_pal The colour palette for the subtitle font. Defaults to "#323232".
-#' @param subtitle_size The size of the subtitle font. Defaults to 10.
-#' @param subtitle_style The style of the subtitle font. Defaults to "plain".
-#' @param body_font The font for all text other than the title, subtitle and caption. If NULL, inherits from font argument.
+#' @param family The font family for all text to use. Defaults to "".
+#' @param title_family The font family for the title. If NULL, inherits from family argument.
+#' @param title_pal The colour palette for the title family. Defaults to "#000000".
+#' @param title_size The size of the title family. Defaults to 11.
+#' @param title_face The font style of the title family. Defaults to "bold".
+#' @param subtitle_family The font family for the subtitle. If NULL, inherits from family argument.
+#' @param subtitle_pal The colour palette for the subtitle family. Defaults to "#323232".
+#' @param subtitle_size The size of the subtitle family. Defaults to 10.
+#' @param subtitle_face The font style of the subtitle family. Defaults to "plain".
+#' @param body_family The font family for all text other than the title, subtitle and caption. If NULL, inherits from family argument.
 #' @param body_pal The colour palette for all text other than the title, subtitle or caption. Defaults to "#323232".
 #' @param body_size The size of all text other than the title, subtitle and caption. Defaults to 10.
-#' @param body_style The style of all text other than the title, subtitle or caption. Defaults to "plain".
-#' @param caption_font The font for the caption. If NULL, inherits from font argument.
+#' @param body_face The font style of all text other than the title, subtitle or caption. Defaults to "plain".
+#' @param caption_family The font family for the caption. If NULL, inherits from family argument.
 #' @param caption_pal The colour palette for the caption. Defaults to "#7F7F7F".
 #' @param caption_size The size of the caption. Defaults to 9.
-#' @param caption_style The style of the caption. Defaults to "plain".
+#' @param caption_face The font style of the caption. Defaults to "plain".
 #' @param axis_pal The colour palette for the axis. Defaults to "#323232".
 #' @param axis_size The size of the axis. Defaults to 0.3.
 #' @param ticks_pal The colour palette for the ticks. Defaults to "#323232".
@@ -36,23 +36,23 @@
 #' @export
 #'
 gg_theme <- function(
-    font = "",
+    family = "",
     title_pal = "#000000",
-    title_font = NULL,
+    title_family = NULL,
     title_size = 11,
-    title_style = "bold",
-    subtitle_font = NULL,
+    title_face = "bold",
+    subtitle_family = NULL,
     subtitle_pal = "#323232",
     subtitle_size = 10,
-    subtitle_style = "plain",
-    body_font = NULL,
+    subtitle_face = "plain",
+    body_family = NULL,
     body_pal = "#323232",
     body_size = 10,
-    body_style = "plain",
-    caption_font = NULL,
+    body_face = "plain",
+    caption_family = NULL,
     caption_pal = "#7F7F7F",
     caption_size = 9,
-    caption_style = "plain",
+    caption_face = "plain",
     axis_size = 0.3,
     axis_pal = "#323232",
     ticks_size = 0.3,
@@ -67,9 +67,9 @@ gg_theme <- function(
     facet_gap_size = 1.5,
     void = FALSE) {
 
-    if (is.null(title_font)) title_font <- font
-    if (is.null(subtitle_font)) subtitle_font <- font
-    if (is.null(body_font)) body_font <- font
+    if (is.null(title_family)) title_family <- family
+    if (is.null(subtitle_family)) subtitle_family <- family
+    if (is.null(body_family)) body_family <- family
 
     if (bg_legend_key_pal == "plot") bg_legend_key_pal <- bg_plot_pal
     else if (bg_legend_key_pal == "panel") bg_legend_key_pal <- bg_panel_pal
@@ -77,17 +77,17 @@ gg_theme <- function(
     if (grid_h == TRUE) {
       if (grid_v == FALSE) { #horizontal
         theme <- ggplot2::theme(
-          text = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style),
-          plot.title = ggplot2::element_text(family = title_font, size = title_size, colour = title_pal, face = title_style, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
-          plot.subtitle = ggplot2::element_text(family = subtitle_font, size = subtitle_size, colour = subtitle_pal, face = subtitle_style, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
-          axis.title.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = body_size / 1.33)),
-          axis.title.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, angle = 90, margin = ggplot2::margin(r = body_size)),
-          legend.title = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0, vjust = 0.5),
-          plot.caption = ggplot2::element_text(family = caption_font, size = caption_size, colour = caption_pal, face = caption_style, hjust = 1, margin = ggplot2::margin(t = caption_size)),
-          axis.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = 2)),
-          axis.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 1, margin = ggplot2::margin(r = 2)),
-          strip.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
-          strip.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
+          text = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face),
+          plot.title = ggplot2::element_text(family = title_family, size = title_size, colour = title_pal, face = title_face, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
+          plot.subtitle = ggplot2::element_text(family = subtitle_family, size = subtitle_size, colour = subtitle_pal, face = subtitle_face, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
+          axis.title.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = body_size / 1.33)),
+          axis.title.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, angle = 90, margin = ggplot2::margin(r = body_size)),
+          legend.title = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0, vjust = 0.5),
+          plot.caption = ggplot2::element_text(family = caption_family, size = caption_size, colour = caption_pal, face = caption_face, hjust = 1, margin = ggplot2::margin(t = caption_size)),
+          axis.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = 2)),
+          axis.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 1, margin = ggplot2::margin(r = 2)),
+          strip.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
+          strip.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
           legend.text = ggplot2::element_text(hjust = 0, margin = ggplot2::margin(r = 7.5, unit = "pt")),
           plot.margin = ggplot2::margin(t = 15, l = 7.5, b = 10, r = 20),
           plot.title.position = "plot",
@@ -118,17 +118,17 @@ gg_theme <- function(
       }
       else if (grid_v == TRUE) { #both
         theme <- ggplot2::theme(
-          text = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style),
-          plot.title = ggplot2::element_text(family = title_font, size = title_size, colour = title_pal, face = title_style, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
-          plot.subtitle = ggplot2::element_text(family = subtitle_font, size = subtitle_size, colour = subtitle_pal, face = subtitle_style, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
-          axis.title.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = body_size / 1.33)),
-          axis.title.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, angle = 90, margin = ggplot2::margin(r = body_size)),
-          legend.title = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0, vjust = 0.5),
-          plot.caption = ggplot2::element_text(family = caption_font, size = caption_size, colour = caption_pal, face = caption_style, hjust = 1, margin = ggplot2::margin(t = caption_size)),
-          axis.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = 2)),
-          axis.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 1, margin = ggplot2::margin(r = 2)),
-          strip.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
-          strip.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
+          text = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face),
+          plot.title = ggplot2::element_text(family = title_family, size = title_size, colour = title_pal, face = title_face, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
+          plot.subtitle = ggplot2::element_text(family = subtitle_family, size = subtitle_size, colour = subtitle_pal, face = subtitle_face, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
+          axis.title.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = body_size / 1.33)),
+          axis.title.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, angle = 90, margin = ggplot2::margin(r = body_size)),
+          legend.title = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0, vjust = 0.5),
+          plot.caption = ggplot2::element_text(family = caption_family, size = caption_size, colour = caption_pal, face = caption_face, hjust = 1, margin = ggplot2::margin(t = caption_size)),
+          axis.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = 2)),
+          axis.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 1, margin = ggplot2::margin(r = 2)),
+          strip.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
+          strip.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
           legend.text = ggplot2::element_text(hjust = 0, margin = ggplot2::margin(r = 7.5, unit = "pt")),
           plot.margin = ggplot2::margin(t = 15, l = 7.5, b = 10, r = 20),
           plot.title.position = "plot",
@@ -160,17 +160,17 @@ gg_theme <- function(
     else if (grid_h == FALSE) {
       if (grid_v == FALSE) { #none
         theme <- ggplot2::theme(
-          text = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style),
-          plot.title = ggplot2::element_text(family = title_font, size = title_size, colour = title_pal, face = title_style, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
-          plot.subtitle = ggplot2::element_text(family = subtitle_font, size = subtitle_size, colour = subtitle_pal, face = subtitle_style, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
-          axis.title.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = body_size / 1.33)),
-          axis.title.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, angle = 90, margin = ggplot2::margin(r = body_size)),
-          legend.title = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0, vjust = 0.5),
-          plot.caption = ggplot2::element_text(family = caption_font, size = caption_size, colour = caption_pal, face = caption_style, hjust = 1, margin = ggplot2::margin(t = caption_size)),
-          axis.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = 2)),
-          axis.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 1, margin = ggplot2::margin(r = 2)),
-          strip.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
-          strip.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
+          text = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face),
+          plot.title = ggplot2::element_text(family = title_family, size = title_size, colour = title_pal, face = title_face, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
+          plot.subtitle = ggplot2::element_text(family = subtitle_family, size = subtitle_size, colour = subtitle_pal, face = subtitle_face, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
+          axis.title.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = body_size / 1.33)),
+          axis.title.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, angle = 90, margin = ggplot2::margin(r = body_size)),
+          legend.title = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0, vjust = 0.5),
+          plot.caption = ggplot2::element_text(family = caption_family, size = caption_size, colour = caption_pal, face = caption_face, hjust = 1, margin = ggplot2::margin(t = caption_size)),
+          axis.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = 2)),
+          axis.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 1, margin = ggplot2::margin(r = 2)),
+          strip.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
+          strip.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
           legend.text = ggplot2::element_text(hjust = 0, margin = ggplot2::margin(r = 7.5, unit = "pt")),
           plot.margin = ggplot2::margin(t = 15, l = 7.5, b = 10, r = 20),
           plot.title.position = "plot",
@@ -201,17 +201,17 @@ gg_theme <- function(
       }
       else if (grid_v == TRUE) { #vertical
         theme <- ggplot2::theme(
-          text = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style),
-          plot.title = ggplot2::element_text(family = title_font, size = title_size, colour = title_pal, face = title_style, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
-          plot.subtitle = ggplot2::element_text(family = subtitle_font, size = subtitle_size, colour = subtitle_pal, face = subtitle_style, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
-          axis.title.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = body_size / 1.33)),
-          axis.title.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, angle = 90, margin = ggplot2::margin(r = body_size)),
-          legend.title = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0, vjust = 0.5),
-          plot.caption = ggplot2::element_text(family = caption_font, size = caption_size, colour = caption_pal, face = caption_style, hjust = 1, margin = ggplot2::margin(t = caption_size)),
-          axis.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, margin = ggplot2::margin(t = 2)),
-          axis.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 1, margin = ggplot2::margin(r = 2)),
-          strip.text.x = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
-          strip.text.y = ggplot2::element_text(family = body_font, size = body_size, colour = body_pal, face = body_style, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
+          text = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face),
+          plot.title = ggplot2::element_text(family = title_family, size = title_size, colour = title_pal, face = title_face, hjust = 0, vjust = body_size / 5, margin = ggplot2::margin(b = body_size)),
+          plot.subtitle = ggplot2::element_text(family = subtitle_family, size = subtitle_size, colour = subtitle_pal, face = subtitle_face, hjust = 0, vjust = body_size / 2.5, margin = ggplot2::margin(t = body_size * -0.2, b = body_size * 0.3)),
+          axis.title.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = body_size / 1.33)),
+          axis.title.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, angle = 90, margin = ggplot2::margin(r = body_size)),
+          legend.title = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0, vjust = 0.5),
+          plot.caption = ggplot2::element_text(family = caption_family, size = caption_size, colour = caption_pal, face = caption_face, hjust = 1, margin = ggplot2::margin(t = caption_size)),
+          axis.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, margin = ggplot2::margin(t = 2)),
+          axis.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 1, margin = ggplot2::margin(r = 2)),
+          strip.text.x = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(b = body_size / 2)),
+          strip.text.y = ggplot2::element_text(family = body_family, size = body_size, colour = body_pal, face = body_face, hjust = 0.5, margin = ggplot2::margin(l = body_size / 2), angle = 270),
           legend.text = ggplot2::element_text(hjust = 0, margin = ggplot2::margin(r = 7.5, unit = "pt")),
           plot.margin = ggplot2::margin(t = 15, l = 7.5, b = 10, r = 20),
           plot.title.position = "plot",